ProblemliMesele

+++ codes/spring/my-spring-project/src/main/java/codes/neriman/my_spring_project/repository/ReaderRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import codes.neriman.my_spring_project.entity.Reader;
 
-public interface ReaderRepository extends JpaRepository<Reader, Long> {
+public interface ReaderRepository extends JpaRepository<Reader, Integer> {
 
 	Optional<Reader> findByUsername(String username);
